feat: move test to streamable_test.go
